create abstract page
